+++ packages/app-democracy/src/Overview/Proposals.tsx
@@ -6,12 +6,10 @@ import { I18nProps } from '@polkadot/ui-app/types';
 
 import React from 'react';
 import { Tuple } from '@polkadot/types';
-import { Button } from '@polkadot/ui-app';
 import { withCalls, withMulti } from '@polkadot/ui-api';
 
 import Proposal from './Proposal';
-import Propose from './Propose';
-import translate from './translate';
+import translate from '../translate';
 
 type Props = I18nProps & {
   democracy_publicProps?: Array<Tuple>
@@ -33,16 +31,8 @@ class Proposals extends React.PureComponent<Props> {
       <section className='democracy--Proposals'>
         <h1>
           {t('proposals')}
-          <Button
-            isPrimary
-            className='democracy--Proposal-new'
-            key='propose'
-            onClick={this.togglePropose}
-            label={t('Propose')}
-          />
         </h1>
         {this.renderProposals()}
-        {this.renderPropose()}
       </section>
     );
   }
@@ -65,23 +55,6 @@ class Proposals extends React.PureComponent<Props> {
         value={proposal}
       />
     ));
-  }
-
-  private renderPropose () {
-    const { isProposeOpen } = this.state;
-
-    return (
-      <Propose
-        isOpen={isProposeOpen}
-        onClose={this.togglePropose}
-      />
-    );
-  }
-
-  private togglePropose = () => {
-    this.setState(({ isProposeOpen }: State) => ({
-      isProposeOpen: !isProposeOpen
-    }));
   }
 }
